requirements.txt
pyodbc

fastapi
pydantic < 1.11.0
python-multipart

asyncio

gunicorn
uvicorn

python-socketio

Jinja2

watchdog